Added UI Stuff
I added these mundane things that our game was in need of:
A main menu that just displays title and closes the game.
A pause menu that has an options, return to main and close the pause menus.
The Options has sound options which dynamically updates via each bus loaded
I rearranged Globals to be a scene instead of just a script and have the ability to play sounds to the new UI Audio Bus.
If you are missing the UI Audio bus, please load it from the file "default_bus_layout.tres"
Also added a set of 8 bit sound effects from a free source, listed in the Credits.txt file

**Important**
I rearranged the Main Menu to be under UI by itself, which is not how the file system diagram is set up, will have to change that.