Updated many things
The base level has decided due to the new change in scope of the project, to be the only level, now supporting a more flat environment and more barren than "green farm". With the introduction of proper unit sprites that fit the sizes, all animations have been changed to match the same sprite sheet. With the addition of these animations, we also have new unit resources. There is now an AI resource, which can handle the application of the enums and presents instead of the GLOBAL VALUES that are preset for EVERY ENEMY UNIT